Issue #4165: Split and Organize Checkstyle inputs by Test for checks in coding package: NestedIfDepth
