+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/nestedifdepth/InputNestedIfDepth.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.coding;
+package com.puppycrawl.tools.checkstyle.checks.coding.nestedifdepth;
 public class InputNestedIfDepth
 {
     void foo() {
